Refactor: check_token_expiration based on provided claims rather than validating token internally for them.
