+++ src/oauth2/token.rs
@@ -68,9 +68,7 @@ use std::time::{Duration, SystemTime, UNIX_EPOCH};
 use jsonwebtoken::{DecodingKey, Validation};
 use log::{debug, error, info, trace};
 use oauth2::basic::BasicTokenType;
-use oauth2::{
-    AccessToken, AuthorizationCode, EmptyExtraTokenFields, RefreshToken, StandardTokenResponse,
-};
+use oauth2::{AuthorizationCode, EmptyExtraTokenFields, RefreshToken, StandardTokenResponse};
 
 use crate::error::{Error, OAuthError};
 use crate::model::oauth2::{EveJwtClaims, EveJwtKey, EveJwtKeys};
@@ -244,84 +242,62 @@ impl<'a> OAuth2Api<'a> {
             }
         }
     }
-
-    /// Checks if the provided access token is expired
-    ///
-    /// Use this method before fetching data from an authenticated ESI route to ensure your access
-    /// token is not expired. This method uses the [`Self::validate_token`] method internally to access
-    /// the claims and ensure the claims have not been modified.
-    ///
-    /// # Arguments
-    /// - `access_token`: An access token in string format. If you haven't converted the
-    ///   token to string yet, you can do so with `token.access_token().secret().to_string()`.
-    ///
-    /// # Returns
-    /// Returns a [`Result`] containing either:
-    /// - [`bool`]: A bool indicating whether or not the token is expired
-    /// - [`Error`]: An error if the token validation used to retrieve the expiration from claims
-    ///   fails. This generally happens if the access token provided is not a valid access
-    ///   token or there is an issue fetching the JWT keys to validate the token.
-    pub async fn check_token_expiration(&self, access_token: &str) -> Result<bool, Error> {
-        let message =
-            "Checking token expiration, attempting to validate token prior to expiration check.";
-        log::debug!("{}", message);
-
-        // Validate token to get claims
-        let access_token = AccessToken::new(access_token.to_string());
-
-        let claims = match self.validate_token(access_token.secret().to_string()).await {
-            Ok(claims) => claims,
-            Err(err) => {
-                // Trace because the validate_token method already logs an error for this
-                let message = format!(
-                    "Failed to validate token for expiration check due to error: {}",
-                    err
-                );
-                log::trace!("{}", message);
-
-                return Err(err);
-            }
-        };
-
-        // Logging: Get character ID for debug logging
-        let id_str = claims.sub.split(':').collect::<Vec<&str>>()[2];
-        let character_id: i64 = id_str.parse().expect("Failed to parse character id to i64");
-
-        // Trace because validate_token already logs info for this
-        let message = format!(
-            "Successfully validated token for character ID {} prior to token expiration check",
-            character_id
-        );
-        log::trace!("{}", message);
-
-        // Check token expiration
-        let expiration_secs = Duration::from_secs(claims.exp as u64);
-        let expiration = UNIX_EPOCH + expiration_secs;
-
-        if SystemTime::now() < expiration {
-            let message = format!(
-                "Checked token for expiration, token for character ID {} is not yet expired, expiration in {}s",
-                character_id,
-                expiration_secs.as_secs()
-            );
-            log::debug!("{}", message);
-
-            // Return false, token is not yet expired
-            return Ok(false);
-        }
-
-        let message = format!(
-            "Checked token for expiration, token for character ID {} is expired",
-            character_id
-        );
-        log::debug!("{}", message);
-
-        // Return true, token is expired
-        Ok(true)
-    }
 }
 
-/// Utility function to validate that the scopes in the provided token claims matches those expected
+/// Utility function to check if token is expired using provided claims
+///
+/// If your token is expired then a request to an authenticated ESI route will return an error. It is ideal to
+/// stop the request from happening on the client side to not incur ESI error limits.
+///
+/// # Arguments
+/// - `claims` ([`EveJwtClaims`]): The claims of a token obtained via the [`Self::validate_token`]
+///   method to be checked if the token is expired.
+///
+/// # Returns
+/// - `bool`: Bool indicating whether or not token is expired
+pub fn check_token_expiration(claims: EveJwtClaims) -> bool {
+    // Logging: Get character ID for debug logging
+    let id_str = claims.sub.split(':').collect::<Vec<&str>>()[2];
+    let character_id: i64 = id_str.parse().expect("Failed to parse character id to i64");
+
+    // Trace because validate_token already logs info for this
+    let message = format!(
+        "Successfully validated token for character ID {} prior to token expiration check",
+        character_id
+    );
+    log::trace!("{}", message);
+
+    // Check token expiration
+    let expiration_secs = Duration::from_secs(claims.exp as u64);
+    let expiration = UNIX_EPOCH + expiration_secs;
+
+    if SystemTime::now() < expiration {
+        // Token is not yet expired
+        let message = format!(
+            "Checked token for expiration, token for character ID {} is not yet expired, expiration in {}s",
+            character_id,
+            expiration_secs.as_secs()
+        );
+        log::debug!("{}", message);
+
+        return false;
+    }
+
+    // Token is expired
+
+    let message = format!(
+        "Checked token for expiration, token for character ID {} is expired",
+        character_id
+    );
+    log::debug!("{}", message);
+
+    true
+}
+
+/// Utility function to check if token has expected scopes
+///
+/// If your token is missing the scopes required for an authenticated ESI route your request will return
+/// an error. It is ideal to stop the request from happening on the client side to not incur ESI error limits.
 ///
 /// # Arguments
 /// - `expected_scopes` (Vec<String>): An array of scope strings validated against the
@@ -330,7 +306,7 @@ impl<'a> OAuth2Api<'a> {
 ///   method to be checked for the expected scopes.
 ///
 /// # Returns
-/// - `bool`: Bool indicating whether or not all expected scopes are present.
+/// - `bool`: true if all expected scopes are present.
 pub fn check_token_scopes(expected_scopes: Vec<String>, claims: EveJwtClaims) -> bool {
     // Logging: Get character ID for debug logging
     let id_str = claims.sub.split(':').collect::<Vec<&str>>()[2];
@@ -513,5 +489,62 @@ mod check_token_scopes_tests {
 
         // Assert result is false
         assert_eq!(result, false);
+    }
+}
+
+#[cfg(test)]
+mod check_token_expiration_tests {
+    use std::time::{SystemTime, UNIX_EPOCH};
+
+    use crate::model::oauth2::EveJwtClaims;
+    use crate::oauth2::token::check_token_expiration;
+
+    /// Test that function returns false due to token not being expired
+    ///
+    /// # Test Setup
+    /// - Create new mock token which is not expired
+    ///
+    /// # Assert
+    /// - Assert result is false
+    #[tokio::test]
+    pub async fn test_check_token_expiration_not_expired() {
+        // Create new mock token which is not expired
+        let mock_claims = EveJwtClaims::mock();
+
+        // Test function
+        let result = check_token_expiration(mock_claims);
+
+        // Assert result is false
+        assert_eq!(result, false);
+    }
+
+    /// Ensures that an expired token returns as expired.
+    ///
+    /// Uses a mock token & mock JWT keys created with a test pair of RSA public & private
+    /// keys. The function under test checks the expiration by validating the mock token to get the
+    /// claims first using the mock JWT keys.
+    ///
+    /// # Test Setup
+    /// - Create new mock token which is expired
+    ///
+    /// # Assert
+    /// - Assert result is true
+    #[tokio::test]
+    async fn test_check_token_expiration_expired() {
+        // Create new mock token which is expired
+        let now = SystemTime::now()
+            .duration_since(UNIX_EPOCH)
+            .unwrap()
+            .as_secs() as i64;
+
+        let mut mock_claims = EveJwtClaims::mock();
+        mock_claims.exp = now - 60;
+        mock_claims.iat = now - 960;
+
+        // Test function
+        let result = check_token_expiration(mock_claims);
+
+        // Assert result is true
+        assert_eq!(result, true);
     }
 }
